requirements.txt
asgiref~=3.2.10
Django~=2.2.16
psycopg2-binary==2.9.5
python-dotenv==1.0.0
sqlparse~=0.3.1

pip~=22.0.2
attrs~=21.2.0
toml~=0.10.2
wheel~=0.37.1
cryptography~=3.4.8
py~=1.11.0
Pillow~=9.0.1
keyring~=23.5.0
colorama~=0.4.4
PyYAML~=5.4.1
requests~=2.25.1
click~=8.0.3
setuptools~=59.6.0
future~=0.18.2
djangorestframework~=3.12.4